MNT: Move TimeParser, FileConverter, and FileHandler to new file_io directory
