+++ trace/tests/test_file_io/test_file_convert.py
@@ -9,10 +9,10 @@ from unittest.mock import Mock, patch
 import pytest
 from qtpy.QtGui import QColor
 
-from trace_file_convert import TraceFileConverter
+from file_io import TraceFileConverter
 
 DUMMY_ARCHIVER_URL = "dummy.archiver.url"
-SCRIPT_PATH = Path(__file__).parent.parent / "trace_file_convert.py"
+SCRIPT_PATH = Path(__file__).parent.parent.parent / "file_io" / "trace_file_convert.py"
 
 
 @pytest.fixture
